add helper method to initiate a direct channel chat
